app/requirements.txt
Django==4.2.9
django-ninja==1.3.0
pytz==2022.4
tzdata==2022.4
django-admin-interface==0.22.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.4
